# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: OneCo Portfolio Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
  /users/me/sync:
    post:
      operationId: syncMe
      tags: [users]
      summary: Sync current user from Clerk (JIT provisioning)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserSync"
      responses:
        "200":
          description: Synced user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserWithProjects"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Anonymize/delete user (admin only, GDPR)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Projects ─────────────────────────────────────────────────────
...
  /projects/{projectId}/tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: assigneeId
          in: query
          schema:
            type: integer
        - name: priority
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create task
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /projects/{projectId}/tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get task
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update task
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete task
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /projects/{projectId}/tasks/{id}/comments:
    get:
      operationId: listTaskComments
      tags: [tasks]
      summary: List comments on a task
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskComment"
    post:
      operationId: createTaskComment
      tags: [tasks]
      summary: Add comment to task
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCommentInput"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskComment"

  # ─── Activity Log ──────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, clerkId, name, email, systemRole, active, createdAt]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        name:
          type: string
        email:
          type: string
        systemRole:
          type: string
          enum: [admin, user]
        active:
          type: boolean
        lastLogin:
          type: ["string", "null"]
        createdAt:
          type: string
    UserWithProjects:
      type: object
      required: [id, clerkId, name, email, systemRole, active, createdAt, projectCount]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        name:
          type: string
        email:
          type: string
        systemRole:
          type: string
          enum: [admin, user]
        active:
          type: boolean
        lastLogin:
          type: ["string", "null"]
        createdAt:
          type: string
        projectCount:
          type: integer
    UserSync:
      type: object
      required: [clerkId, name, email]
      properties:
        clerkId:
          type: string
        name:
          type: string
        email:
          type: string
    UserInput:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
        systemRole:
          type: string
          enum: [admin, user]
    UserUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
        systemRole:
          type: string
          enum: [admin, user]
        active:
          type: boolean

    # ─── Project Schemas ─────────────────────────────────────────────
...
    Task:
      type: object
      required: [id, projectId, title, status, priority, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        assigneeId:
          type: ["integer", "null"]
        assigneeName:
          type: ["string", "null"]
        status:
          type: string
          enum: [ikke_startet, pagaar, venter, fullfort]
        priority:
          type: string
          enum: [lav, middels, hoy]
        dueDate:
          type: ["string", "null"]
        createdAt:
          type: string
...
    TaskInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        assigneeId:
          type: integer
        status:
          type: string
          enum: [ikke_startet, pagaar, venter, fullfort]
        priority:
          type: string
          enum: [lav, middels, hoy]
        dueDate:
          type: string
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        assigneeId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [ikke_startet, pagaar, venter, fullfort]
        priority:
          type: string
          enum: [lav, middels, hoy]
        dueDate:
          type: ["string", "null"]
    TaskComment:
      type: object
      required: [id, taskId, userId, text, createdAt]
      properties:
        id:
          type: integer
        taskId:
          type: integer
        userId:
          type: integer
        userName:
          type: ["string", "null"]
        text:
          type: string
        createdAt:
          type: string
    TaskCommentInput:
      type: object
      required: [text]
      properties:
        text:
          type: string
          minLength: 1

    # ─── Activity Schemas ────────────────────────────────────────────
...